Mongo poc with file implementation

+++ src/Blogifier.Shared/Models/AnalyticsModel.cs
@@ -2,10 +2,10 @@ namespace Blogifier.Shared
 {
 	public class AnalyticsModel
 	{
-		public int TotalPosts { get; set; }
-        public int TotalPages { get; set; }
-		public int TotalViews { get; set; }
-		public int TotalSubscribers { get; set; }
+		public long TotalPosts { get; set; }
+        public long TotalPages { get; set; }
+		public long TotalViews { get; set; }
+		public long TotalSubscribers { get; set; }
 
         public AnalyticsListType DisplayType { get; set; }
         public AnalyticsPeriod DisplayPeriod { get; set; }
